+++ src/MS_gt_07.cpp
@@ -13,8 +13,8 @@
  */
 double MS_gt_07::CalculateMassLossRateHurley() {
     double rateNJ = CalculateMassLossRateNieuwenhuijzenDeJager();
+    m_DominantMassLossRate = MASS_LOSS_TYPE::GB;
     if (utils::Compare(rateNJ, 0.0) > 0) {
-        m_DominantMassLossRate = MASS_LOSS_TYPE::NIEUWENHUIJZEN_DE_JAGER;
     } else {
         m_DominantMassLossRate = MASS_LOSS_TYPE::NONE;
     }